requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
sqlalchemy==2.0.40
pydantic==2.11.4
python-jose==3.4.0
python-multipart==0.0.20
passlib==1.7.4
bcrypt==4.3.0
email-validator==2.1.1 